Rethought notebook events & reactivity
The standard way that jupyter connects to notebooks is not great for our purposes. The heedy notebook needs to auto-update for many users, and should only allow executing cell code. Also, heedy already contains an event system, there is not reason to open another websocket.

Finally, doesn't make sense to re-send the entire notebook to save it.

With that in mind, notebooks are now held in the database, and an update-based saving mechanism is implemented. This allows notebooks to auto-update on save for multiple viewers. Furthermore, executing notebook cells is currently disconnected, but will be done completely in the backend. Heedy's events will notify frontends when new data is available for a given cell, so no need for any new sockets.